Add support for client authenticated TLS.
This adds three new optional command line arguments to the client:

clientcert = PEM encoded client certificate
clientkey = PEM encoded client key
cacert = PEM encoded trusted root CA certificate

The clientcert and clientkey parameters must both be passed to enable
client authentication of a TLS connection.  The cacert parameter is
optional and is only needed if the server being connected to is signed
by a CA that is not trusted by default.

If the arguments are left off the default TLS config will be used (no
change in existing behavior).